+++ doc/release_notes.rst
@@ -14,6 +14,8 @@ Upcoming Release
 PyPSA 0.18.1 (15th October 2021)
 ================================
 
+* Add assert: CBC solver does not work with '>' and '<'
+
 * Compatibility with ``pyomo>=6.1``.
 
 * Bugfix: specifying the ``solver_logfile`` is no longer mandatory with CPLEX for